add try-catch
