+++ controllers/contactsController.js
@@ -1,0 +1,109 @@
+const {
+  listContacts,
+  getContactById,
+  removeContact,
+  addContact,
+  updateContact,
+} = require("../models/contacts");
+
+const getContacts = async (req, res, next) => {
+  try {
+    const contacts = await listContacts();
+    res.status(200).json({
+      status: "success",
+      code: 200,
+      contacts,
+    });
+  } catch (error) {
+    next(error);
+  }
+};
+
+const getOneContact = async (req, res, next) => {
+  try {
+    const { id } = req.params;
+    const contact = await getContactById(id);
+
+    if (!contact) {
+      return res.status(404).json({
+        status: "error",
+        code: 404,
+        message: "Not found",
+      });
+    }
+    res.status(200).json({
+      status: "success",
+      code: 200,
+      contact,
+    });
+  } catch (error) {
+    next(error);
+  }
+};
+
+const deleteContact = async (req, res, next) => {
+  try {
+    const { id } = req.params;
+    const contact = await removeContact(id);
+    console.log(contact);
+    if (!contact) {
+      return res.status(404).json({
+        status: "error",
+        code: 404,
+        message: "Not found",
+      });
+    }
+    res.status(200).json({
+      status: "success",
+      code: 200,
+      message: "contact deleted",
+      data: {
+        contact,
+      },
+    });
+  } catch (error) {
+    next(error);
+  }
+};
+
+const postContact = async (req, res, next) => {
+  try {
+    const contact = await addContact(req.body);
+
+    res.status(201).json({
+      status: "success",
+      code: 201,
+      data: {
+        contact,
+      },
+    });
+  } catch (error) {
+    next(error);
+  }
+};
+
+const putContact = async (req, res, next) => {
+  try {
+    const { id } = req.params;
+
+    const contact = await updateContact(id, req.body);
+
+    res.status(200).json({
+      status: "success",
+      code: 200,
+      data: {
+        contact,
+      },
+    });
+  } catch (error) {
+    next(error);
+  }
+};
+
+module.exports = {
+  getContacts,
+  getOneContact,
+  deleteContact,
+  postContact,
+  putContact,
+};
