+++ middlewares/validationMiddleware.js
@@ -1,0 +1,21 @@
+const Joi = require("joi");
+const postValidation = (req, res, next) => {
+  const schema = Joi.object({
+    name: Joi.string().min(3).max(30).required(),
+    phone: Joi.string().required(),
+    email: Joi.string().email().required(),
+  });
+  const validationResult = schema.validate(req.body);
+  if (validationResult.error) {
+    return res.status(400).json({
+      status: "error",
+      code: 400,
+      message: "missing fields",
+    });
+  }
+  next();
+};
+
+module.exports = {
+  postValidation,
+};
