+++ models/contacts.json
@@ -54,9 +54,15 @@
     "phone": "(956) 914-3792"
   },
   {
-    "id": "b1SNgO5M3F1opF1ewBe9T",
-    "name": "Anna Lewis",
-    "email": "anna.in@egetlacus.ca",
-    "phone": "(299) 840-6685"
+    "id": "Sz2eGYb_ploDnp2lTp5_A",
+    "name": "Rostyk",
+    "email": "spirit@gmail.com",
+    "phone": "(5999) 472-5218"
+  },
+  {
+    "id": "43fXwOu0opVYupMQAD0NA",
+    "name": "mama",
+    "email": "mama@gmail.com",
+    "phone": "(5999) 472-5218"
   }
 ]
